dev merge

+++ src/main/java/dev/yerokha/flightstatusapi/application/dto/CustomPage.java
@@ -4,11 +4,12 @@ import jakarta.validation.constraints.NotNull;
 
 import java.util.List;
 
+
 public record CustomPage<T>(
         @NotNull List<T> content,
         @NotNull int pageNumber,
         @NotNull long pageSize,
-        @NotNull int totalElements,
+        @NotNull long totalElements,
         @NotNull int totalPages,
         @NotNull boolean isEmpty
 ) {
